Added mysql sql client command builder for issue #372
